fix(editor): unable to move word left/right on a single line
!darwin:
Moving by word gets hung up in (left or right arrow) gets hung up in the
check for ModShorcut. (copy/paste/etc). In this case we should just check
if the modifer comes along with an arrow key.

This also add checking for _Unicode Punctuation_ instead of just spaces
when moving a word.

Cheers,
-Nate